refactor(WishProduct): 패키지명 변경
