+++ src/main/java/shopping/wishproduct/application/WishProductRepository.java
@@ -1,9 +1,9 @@
-package shopping.wishlist.application;
+package shopping.wishproduct.application;
 
 import java.util.List;
 import shopping.member.domain.Member;
 import shopping.product.domain.Product;
-import shopping.wishlist.domain.WishProduct;
+import shopping.wishproduct.domain.WishProduct;
 
 public interface WishProductRepository {
     boolean existsByMemberAndProduct(Member member, Product product);
