+++ src/main/java/shopping/wishproduct/infra/WishProductRepositoryImpl.java
@@ -1,12 +1,12 @@
-package shopping.wishlist.infra;
+package shopping.wishproduct.infra;
 
 import java.util.List;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import shopping.member.domain.Member;
 import shopping.product.domain.Product;
-import shopping.wishlist.application.WishProductRepository;
-import shopping.wishlist.domain.WishProduct;
+import shopping.wishproduct.application.WishProductRepository;
+import shopping.wishproduct.domain.WishProduct;
 
 @Repository
 public class WishProductRepositoryImpl implements WishProductRepository {
